Add devise omniauthable and confirmable to User
Rolled back the database and uncommented the omniauthable and
confirmable sections in the devise create users migration file. Also
uncommented the relevant sections in the User model.

Added the omniauth-github gem to the Gemfile.

Generated a new migration to add provider and uid columns to the users
table.

Migrated the database.