Notification system implementation 1
Renamed notification_obj_refs to notification_objects, and renamed the
polymorphic interface to notification_triggerable because I felt it
improved the readability of the models.

Implemented a before filter in NotificationChange to set the description
of the change based on the notification_triggerable_type.

Implemented an after filter in FriendRequest to set up all of the
necessary records for user notification upon request creation.

Tested the filters in the FriendRequest test, confirmed all tests pass.