Implement redesigned site map.
Created a splash page and a controller for static pages. In the
controller home method, if the user is not logged in, the splash page
will be rendered. If the user is signed in and has a location set, they
are redirected to their feed.

Modified after_sign_in_path_for(resource) to direct the user to
locations#new if their location is nil upon sign in.

Wrote tests for new site map, confirmed all tests pass.

Tested Location model and dependent destroy in user.