Implement location_set? in ApplicationController
Wanted a filter to check that the user's location is set after they sign
in. That way, if a user signs in, then types in a URL, they can't access
it without setting their location first. No matter what, the user has to
visit the new_location_path, but will be able to decline and set their
location to secret.