Move location_set? filter, Implement working location selection
Location_set? filter was causing a redirect loop because it was being
called by the LocationsController. Thus, I moved the filer callback out
of the ApplicationController and into the individual controllers that
needed it.

Added a notice about the use of location services to the locations#new
view. The user can accept, which dismisses the notice, or can decline,
which sets the user's location to The Bermuda Triangle.

Added the necessary logic to the locations_base.js.erb file to set the
values of the info window form to the place title, and lat/lng. When the
user accepts the location, it creates a location object with that
information belonging to the user.