Fix bugs that disabled a user's ability to search multiple locations
The _locations_base.js.erb file was written incorrectly and kept a user
from searching multiple locations. The user could pull up a location,
click it and see the info window only once. I had to add code to clear
the previous infowindows, clone the hidden form, and edit the form
values after clicking on a marker. Now a user can click a location,
decide against using it, make a new search, and click on a new marker to
submit the form. Only one infowindow will open at a time.

Added instructions to the user to tell them not to select their full
address, so that they are still mostly anonymous in their location.

Refactored locations#new.