Implement user tagging system edit action and view
Tags can be selected and added to a form to update the user, but a
couple of things need to be done before moving on to update.

The user should not be able to select the same tag more than once.

The search box needs to be cleared after selecting a tag.

The form styling needs to be completed and tag buttons need to be
styled.