Implement UserLocation model to associate a user with one location
Realized the way I handled locations originally was incorrect.
Every time a user set their location, a new location was created, even
so that the user can opt to disable their location but no location will
ever be nil and handling users without a location will be possible.

I decided to enforce some null constraints on the locations table so
that a location without title, latitude, or longitude could not exist. I
I decided to enforce some not null constraints on the locations table so
that a location without title, latitude, or longitude can't exist. I
also added indices to enforce the uniquness of latitude and longitude as
a pair, to prevent multiples of the same location from being created. I
also added a uniqueness constraint on the title of a location so that
the same title cannot be used more than once. I want this to be used to
pressure the user to enter more specific information about their
location (Stafford, VA instead of just Stafford, since the town of
Stafford may exist all over in different localities). I will add a
validation for this in the model next.

Revised tests that created multiple users using create(:confirmed_user).
The confirmed_user factory creates a location for each user, but the
locations were not guaranteed to be unique, so some tests were throwing
errors because of uniqueness constraints in the database indices. Thus,
it was necessary to change the way that users were being created in any
test where multiple users were being created, in order to ensure that
tests will always pass.

Created a UserLocationsController to handle setting, updating, and
disabling user_locations. Methods were taken from the
LocationsController and modified where necessary to pass the
corresponding tests I created. Also created routes for this controller.
I will delete the routes for the location controller but leave the
controller in place so in the future the controller could be used by
admins to manipulate existing locations.