Preliminary implementation of Users#edit_skill_list
Added require_skills to ApplicationController. This method will redirect
the user to set their skills if they haven't already. It is used in all
controllers except for the static pages controller. As such, all a new
user factory was created and tests were added for this method in all
controller tests.

Renamed the user factories after the addition of the user factory with
skills. It was necessary to fix tests that were affected by the naming
changes afterwards. Refactored some of the tests as well.

Implemented methods in FriendRequests, Friendships, and Notifications
controllers to prevent the wrong user from accessing the destroy action.

Modified redirect path in UserLocations#create to send the user to set
their skills. Reflected this change in the site map tests.