Update skill tag views and script
Script will now populate the edit_skill_list view with tags if the
user's skill list is not empty. There will be no new or create actions,
only this one so it needs to be flexible enough to work if a user has
skills or not.

Script will now disable the submit button if a user has not selected any
tags.

Refactored script.

Began testing edit_skill_list.